refactor(BcaTextField): move component to its own directory
