+++ src/components/drawers/Settings/Projects/ProjectDrawer.tsx
@@ -7,7 +7,7 @@ import { zodResolver } from '@hookform/resolvers/zod'
 import DrawerTitle from '~components/titles/DrawerTitle'
 import BcaDrawer from '~components/drawers/BcaDrawer/BcaDrawer'
 import { projectSchema, ProjectType } from '~types/project'
-import BcaTextField from '~components/input/BcaTextField'
+import BcaTextField from '~/components/input/BcaTextField/BcaTextField'
 import ButtonGroup from '~components/buttons/button-group'
 import {
     useCreateProjectMutation,
